+++ contracts/interfaces/IBurnieCoinflip.sol
@@ -92,7 +92,7 @@ interface IBurnieCoinflip {
     function processCoinflipPayouts(
         bool bonusFlip,
         uint256 rngWord,
-        uint32 epoch
+        uint24 epoch
     ) external;
 
     /*+======================================================================+
@@ -150,7 +150,7 @@ interface IBurnieCoinflip {
             bool enabled,
             uint256 stop,
             uint256 carry,
-            uint32 startDay
+            uint24 startDay
         );
 
     /// @notice Get the top bettor from the most recently resolved flip day.
@@ -175,5 +175,5 @@ interface IBurnieCoinflip {
     /// @param day The day to query.
     /// @return rewardPercent The reward percentage for that day.
     /// @return win Whether the flip was a win.
-    function getCoinflipDayResult(uint32 day) external view returns (uint16 rewardPercent, bool win);
+    function getCoinflipDayResult(uint24 day) external view returns (uint16 rewardPercent, bool win);
 }
